+++ domain/src/main/kotlin/team/retum/domain/param/company/FetchCompaniesParam.kt
@@ -1,6 +1,6 @@
 package team.retum.domain.param.company
 
 data class FetchCompaniesParam(
-    val page: Int,
+    val page: Long,
     val name: String?,
 )
