+++ domain/src/main/kotlin/team/retum/domain/param/recruitment/FetchRecruitmentsParam.kt
@@ -1,7 +1,7 @@
 package team.retum.domain.param.recruitment
 
 data class FetchRecruitmentsParam(
-    val page: Int,
+    val page: Long,
     val name: String?,
     val jobCode: Long?,
     val techCode: String?,
